Move activity-routes-lazyload.js content to activity-routes.js
