+++ legacy/app/activity/activity-routes.js
@@ -1,3 +1,26 @@
 angular.module('ushahidi.activity.routes', [])
 
-.config(require('./activity-routes-lazyload.js'));
+.config([
+    '$stateProvider',
+    '$urlMatcherFactoryProvider',
+    function (
+        $stateProvider,
+        $urlMatcherFactoryProvider
+    ) {
+        $urlMatcherFactoryProvider.strictMode(false);
+
+        $stateProvider
+        .state({
+            name: 'activity',
+            url: '/activity',
+            controller: require('./activity.controller.js'),
+            template: require('./activity.html'),
+            lazyLoad: function ($transition$) {
+                const $ocLazyLoad = $transition$.injector().get('$ocLazyLoad');
+                return System.import('/activity.js').then(mod => {
+                    $ocLazyLoad.load(mod.ACTIVITY_MODULE);
+                });
+            }
+        });
+    }]
+);
